requirements.txt
addict==2.4.0
apex==0.9.10dev
basicsr==1.4.2
cityscapesscripts==2.2.4
decord==0.6.0
diffusers==0.16.0
einops==0.8.1
fairscale==0.4.13
flash_attn==2.8.3
ftfy==6.3.1
horovod==0.28.1
huggingface_hub==0.25.2
iopath==0.1.10
ipython==8.12.3
kornia==0.8.1
lmdb==1.7.3
matplotlib==3.10.5
mc==0.0
numpy==2.3.2
omegaconf==2.3.0
onnx==1.19.0
openai==1.102.0
opencv_python==4.11.0.86
opendatasets==0.1.22
packaging==25.0
pandas==2.3.2
parrots==1.0.3
peft==0.17.0
Pillow==11.3.0
prettytable==3.16.0
pycocoevalcap==1.2
pycocotools==2.0.10
python_magic==0.4.27
PyYAML==6.0.2
PyYAML==6.0.2
regex==2024.11.6
Requests==2.32.5
scikit_learn==1.2.2
scipy==1.16.1
shortuuid==1.0.13
skimage==0.0
spacy==3.8.7
tensorboardX==2.6.4
tensorflow==2.20.0
tifffile==2025.6.11
tiktoken==0.11.0
timm==0.6.13
torch==2.0.1
torch_tb_profiler==0.4.3
torchvision==0.15.2
tqdm==4.67.1
transformers==4.31.0
transformers_stream_generator==0.0.5
turbojpeg==0.0.2
webdataset==0.2.100
yapf==0.43.0
